+++ changelog.md
@@ -15,3 +15,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Setup Users page
 * Setup Libraries page
 * Setup Roles page
+* Setup User Groups page